requirements.txt
#
# Common requirements for wger
#

# Application
Django==4.2.6
bleach[css]~=6.1
celery[redis]==5.3.6
crispy-bootstrap5==2023.10
django-activity-stream==2.0.0
django-axes[ipware]==6.3.0
django-bootstrap-breadcrumbs2==1.0.0 # fork of django-bootstrap-breadcrumbs, we might need to migrate away completely
django-crispy-forms~=2.1
django-email-verification~=0.3.3
django-environ==0.11.2
django-formtools~=2.5
django-recaptcha~=4.0.0
django-simple-history~=3.4
django-storages~=1.14
django_compressor~=4.4
django_extensions~=3.2
drf-spectacular[sidecar]==0.27.1
easy-thumbnails==2.8.5
flower==2.0.1
fontawesomefree~=6.5.1
icalendar==5.0.11
invoke==2.2.0
openfoodfacts==0.1.12
pillow==10.2.0
reportlab==4.0.9
requests==2.31.0
tqdm==4.66.1
tzdata==2023.4

# AWS
#boto3

# REST API
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework~=3.14
djangorestframework-simplejwt[crypto]==5.3.1

# Not used anymore, but needed because some modules are imported in DB migration
# files
django-sortedm2m~=3.1
